requirements.txt
pandas
numpy
nptyping
requests
caveclient
pyyaml
slack_bolt